requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
django-contrib-comments==1.9.1
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
libmagic==1.0
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.2
python-magic-bin==0.4.14
pytz==2019.3
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
uuid==1.30
wincertstore==0.2
wrapt==1.11.2
